% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_synonymous_codons.R
\name{find_synonymous_codons}
\alias{find_synonymous_codons}
\title{find_synonymous_codons}
\usage{
find_synonymous_codons(selectedcodon = NULL, codon_table = codon_table,
  returndf = TRUE, verbose = TRUE)
}
\arguments{
\item{selectedcodon}{codon as string, eg "AAA", to be used as the starting
codon to find synonyms for.}

\item{codon_table}{dataframe object to be used as lookup table. Defaults to
object named codon_table.}

\item{returndf}{logical. Should returned object be a dataframe (or if not,
then list)? Defaults to TRUE.}

\item{verbose}{logical. Defaults to TRUE.}
}
\description{
Takes as input a codon and a codon lookup table, find its encoded amino acid,
and finds alternative codons for that amino acid and returns them as
dataframe (default) or list. Verbose option helps to follow the function as
it is working. Is called by synonymise function, but can also be used
independently.
}
\examples{
synonymous_codons <- find_synonymous_codons(selectedcodon = "CGC", codon_table = codon_table, returndf = TRUE, verbose = FALSE)
}
