% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons.R
\name{loadrarecodons}
\alias{loadrarecodons}
\title{loadrarecodons}
\usage{
loadrarecodons()
}
\description{
Loads a list of rare codons (in E. coli K12). Doesn't
include stop codons. From
http://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=83333.
}
\examples{
rare_codons <- loadrarecodons()
}
