% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons_B.R
\name{loadrarecodons_B}
\alias{loadrarecodons_B}
\title{loadrarecodons_B}
\usage{
loadrarecodons_B()
}
\description{
Loads a list of rare codons (in E. coli B strain). From
https://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=413997.
}
\examples{
rare_codons_B <- loadrarecodons_B()
}
