% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons_Ecoli.R
\name{loadrarecodons_Ecoli}
\alias{loadrarecodons_Ecoli}
\title{loadrarecodons_Ecoli}
\usage{
loadrarecodons_Ecoli()
}
\description{
Loads a list of rare codons (in E. coli K12/B strain).
}
\examples{
rare_codons_Ecoli <- loadrarecodons_Ecoli()
}
