% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons_K12.R
\name{loadrarecodons_K12}
\alias{loadrarecodons_K12}
\title{loadrarecodons_K12}
\usage{
loadrarecodons_K12()
}
\description{
Loads a list of rare codons (in E. coli K12 strain). From
https://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=83333
}
\examples{
rare_codons_K12 <- loadrarecodons_K12()
}
