% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons_B.R
\name{loadveryrarecodons_B}
\alias{loadveryrarecodons_B}
\title{loadveryrarecodons_B}
\usage{
loadveryrarecodons_B()
}
\description{
Loads a list of very rare codons (in E. coli B strain). Doesn't include stop
codons. From
https://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=413997.
}
\examples{
very_rare_codons_B <- loadveryrarecodons_B()
}
