% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons_Ecoli.R
\name{loadveryrarecodons_Ecoli}
\alias{loadveryrarecodons_Ecoli}
\title{loadveryrarecodons_Ecoli}
\usage{
loadveryrarecodons_Ecoli()
}
\description{
Loads a list of very rare codons (in either E. coli K12/B strain). Doesn't include stop
codons.
}
\examples{
very_rare_codons_Ecoli <- loadveryrarecodons_Ecoli()
}
