% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons_K12.R
\name{loadveryrarecodons_K12}
\alias{loadveryrarecodons_K12}
\title{loadveryrarecodons_K12}
\usage{
loadveryrarecodons_K12()
}
\description{
Loads a list of very rare codons (in E. coli K12 strain). Doesn't include stop
codons. From
https://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=83333
}
\examples{
very_rare_codons_K12 <- loadveryrarecodons_K12()
}
