% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_alternative.R
\name{select_alternative}
\alias{select_alternative}
\title{select_alternative}
\usage{
select_alternative(possible_codons, forbidden_codons, verbose = FALSE,
  input = "")
}
\arguments{
\item{possible_codons}{dataframe of codons to choose from (eg. output from
find_synonymous_codons function).}

\item{forbidden_codons}{dataframe of codons to exclude, such as rare codons.}

\item{verbose}{logical. Defaults to FALSE.}

\item{input}{optional. string used to specify input codon, only required for
verbosity. Defaults to "".}
}
\description{
Select a codon from a provided list that has a decent codon usage. Works by
selecting a codon at random from the possible_codons list until it isn't also
on the excluded forbidden_codons list. Codon lists with only one variant, or
with no non-forbidden options will return original codon. Returned codon is
provided as lower case string. Verbose option helps to follow the function as
it is working. Is called by synonymise function, but can also be used
independently.
}
\examples{
select_alternative(possible_codons = synonymous_codons, forbidden_codons = c(very_rare_codons, rare_codons), verbose=TRUE, input="CGC")
}
