% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcodons.R
\name{loadveryrarecodons}
\alias{loadveryrarecodons}
\title{loadveryrarecodons}
\usage{
loadveryrarecodons()
}
\description{
Loads a list of very rare codons (in E. coli K12). Doesn't include stop
codons. From
http://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=83333.
}
\examples{
very_rare_codons <- loadveryrarecodons()
}
